cleans some stuff up
